move sample to its own file
This certainly must go under `simulations`, but it is composed of a sufficiently different functionality to be in a different file.
