+++ src/Agents.jl
@@ -7,8 +7,6 @@ using Random
 import Base.Iterators.product
 import Base.iterate
 import Base.length
-import StatsBase.sample
-import StatsBase.Weights
 
 include("core/model.jl")
 include("core/space.jl")
@@ -16,6 +14,7 @@ include("core/agent_space_interaction.jl")
 include("simulations/data_collector.jl")
 include("simulations/step.jl")
 include("simulations/parallel.jl")
+include("simulations/sample.jl")
 include("CA1D.jl")
 include("CA2D.jl")
